Fix flaky tests in resolver.test.js
There were two flaky tests, both related to assertions on specific
orderings of arrays. One case expected the order of `user.task` to be
deterministic which is not the case, the other expected the order of
the graphql schemas `users` field response to be deterministic.

The issues were fixed by just asserting that members of the compared
arrays were deep equal.