Allow functions for exclude and only options (#289)
This modifies the `exclude` and `only` options of attributeFields to allow
functions to be passed instead of a list of arguments.

This allows more programmatic removal of fields. For example, I would like to
remove all fields that have an underscore in them (foreign keys). This would
allow me to write:

```js
{
  fields: () => ({
    ...attributeFields(Model, { exclude: field => ~field.indexOf('_') })
  })
}
```