# This is a combination of 2 commits. # The first commit's message is:
Fix connection pagination

If you were to query first: 5, then use the resulting cursor
to query last: 2, before: cursor you would actually get a wrong value
because the order would be reversed on you to start with and the resulting
index would be entirely wrong.

This commit only reverses the order for the initial last query which is only
to support literally paginating from the very back of the results.

If before or after are included it is assumed that you started from somewhere and
the order of the query is not reversed.

To facilitate this we make sure we have the full count and an inital query
with only last provided starts counting from fullCount backwards to get proper
indexes in relation to a non reversed query.

More details to be provided in the PR

# This is the 2nd commit message:

Linting and style update

Just some simple linting and style updates to how things were written
functionality remains the same.